refactor: 사용하지 않는 레거시 코드 삭제
