+++ src/main/java/roomescape/controller/AdminController.java
@@ -10,22 +10,22 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.AdminReservationRequest;
-import roomescape.dto.ReservationResponse;
-import roomescape.service.ReservationServiceV2;
+import roomescape.dto.request.AdminReservationRequest;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.service.ReservationService;
 
 @RestController
 public class AdminController {
 
-    private final ReservationServiceV2 reservationServiceV2;
+    private final ReservationService reservationService;
 
-    public AdminController(final ReservationServiceV2 reservationServiceV2) {
-        this.reservationServiceV2 = reservationServiceV2;
+    public AdminController(final ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @PostMapping("/admin/reservations")
     public ResponseEntity<ReservationResponse> addReservation(@RequestBody @Valid final AdminReservationRequest request) {
-        ReservationResponse response = reservationServiceV2.addReservationForAdmin(request);
+        ReservationResponse response = reservationService.addReservationForAdmin(request);
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
@@ -36,6 +36,6 @@ public class AdminController {
             @RequestParam(required = false, name = "dateFrom") LocalDate dateFrom,
             @RequestParam(required = false, name = "dateTo") LocalDate dateTo
     ) {
-        return ResponseEntity.status(HttpStatus.OK).body( reservationServiceV2.getFilteredReservations(memberId, themeId, dateFrom, dateTo));
+        return ResponseEntity.status(HttpStatus.OK).body( reservationService.getFilteredReservations(memberId, themeId, dateFrom, dateTo));
     }
 }
