+++ src/main/java/roomescape/dto/request/AdminReservationRequest.java
@@ -1,9 +1,10 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import java.time.LocalDate;
 import org.springframework.format.annotation.DateTimeFormat;
 
-public record ReservationRequestV2(
+public record AdminReservationRequest(
+        long memberId,
         @DateTimeFormat(pattern = "yyyy-MM-dd")
         LocalDate date,
         long themeId,
