+++ src/main/java/roomescape/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 public record LoginRequest(
         String email,
