+++ src/main/java/roomescape/dto/request/MemberRegisterRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 public record MemberRegisterRequest(
         String email,
