+++ src/main/java/roomescape/dto/request/ReservationRequest.java
@@ -1,12 +1,12 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import java.time.LocalDate;
 import org.springframework.format.annotation.DateTimeFormat;
 
 public record ReservationRequest(
-        String name,
         @DateTimeFormat(pattern = "yyyy-MM-dd")
         LocalDate date,
         long themeId,
-        long timeId) {
+        long timeId
+) {
 }
