+++ src/main/java/roomescape/dto/request/ReservationThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 public record ReservationThemeRequest(String name, String description, String thumbnail) {
 }
