+++ src/main/java/roomescape/dto/response/LoginResponse.java
@@ -1,0 +1,4 @@
+package roomescape.dto.response;
+
+public record LoginResponse(String name) {
+}
