+++ src/main/java/roomescape/dto/response/ReservationResponse.java
@@ -1,8 +1,7 @@
-package roomescape.dto;
+package roomescape.dto.response;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationV2;
 
 public record ReservationResponse(
         long id,
@@ -12,12 +11,12 @@ public record ReservationResponse(
         ReservationThemeResponse theme) {
 
     public static ReservationResponse from(Reservation reservation) {
-        return new ReservationResponse(reservation.getId(), reservation.getName(),
+        return new ReservationResponse(reservation.getId(), reservation.getMember().getName(),
                 reservation.getDate(), ReservationTimeResponse.from(reservation.getTime()),
                 ReservationThemeResponse.from(reservation.getTheme()));
     }
 
-    public static ReservationResponse fromV2(ReservationV2 reservation) {
+    public static ReservationResponse fromV2(Reservation reservation) {
         return new ReservationResponse(reservation.getId(), reservation.getMember().getName(), reservation.getDate(),
                 ReservationTimeResponse.from(reservation.getTime()),
                 ReservationThemeResponse.from(reservation.getTheme()));
