+++ src/main/java/roomescape/dto/response/ReservationThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.response;
 
 import roomescape.domain.ReservationTheme;
