+++ src/main/java/roomescape/repository/jpa/ReservationJpaRepository.java
@@ -4,11 +4,11 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationV2;
+import roomescape.domain.Reservation;
 
 @Repository
-public interface ReservationJpaRepository extends JpaRepository<ReservationV2, Long> {
+public interface ReservationJpaRepository extends JpaRepository<Reservation, Long> {
     boolean existsByDateAndTimeIdAndThemeId(LocalDate date, long timeId, long themeId);
 
-    List<ReservationV2> findByMemberId(Long memberId);
+    List<Reservation> findByMemberId(Long memberId);
 }
