+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,13 +1,18 @@
 package roomescape.service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
+import roomescape.domain.Member;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTheme;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
+import roomescape.dto.request.AdminReservationRequest;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.response.MyPageReservationResponse;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationThemeRepository;
 import roomescape.repository.ReservationTimeRepository;
@@ -15,53 +20,90 @@ import roomescape.repository.ReservationTimeRepository;
 @Service
 public class ReservationService {
 
-    public static final int DELETE_FAILED_COUNT = 0;
-
     private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
     private final ReservationThemeRepository reservationThemeRepository;
+    private final MemberRepository memberRepository;
 
     public ReservationService(final ReservationRepository reservationRepository,
                               final ReservationTimeRepository reservationTimeRepository,
-                              final ReservationThemeRepository reservationThemeRepository) {
+                              final ReservationThemeRepository reservationThemeRepository,
+                              final MemberRepository memberRepository) {
         this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
         this.reservationThemeRepository = reservationThemeRepository;
+        this.memberRepository = memberRepository;
     }
 
-    public List<ReservationResponse> findReservations() {
-        List<Reservation> reservations = reservationRepository.findAll();
-        return reservations.stream().map(ReservationResponse::from).toList();
-    }
-
-    public ReservationResponse addReservation(final ReservationRequest request) {
-        long timeId = request.timeId();
+    public ReservationResponse addReservationWithMemberId(final ReservationRequest request, final long memberId) {
+        final long timeId = request.timeId();
         final long themeId = request.themeId();
-
-        ReservationTime time = reservationTimeRepository.findById(timeId).orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 예약 시간 입니다."));
-        ReservationTheme theme = reservationThemeRepository.findById(themeId).orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 테마 입니다."));
-        Reservation reservation = new Reservation(request.name(), request.date(), time, theme);
-        validateUniqueReservation(reservation);
-        Reservation saved = reservationRepository.save(reservation);
-        return ReservationResponse.from(saved);
+        final Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 사용자 입니다."));
+        final ReservationTime time = reservationTimeRepository.findById(timeId)
+                .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 예약 시간 입니다."));
+        final ReservationTheme theme = reservationThemeRepository.findById(themeId)
+                .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 테마 입니다."));
+        validateDuplicateReservation(request.date(), timeId, themeId);
+        final Reservation reservation = new Reservation(member, request.date(), time, theme);
+        Reservation saved = reservationRepository.saveWithMember(reservation);
+        return ReservationResponse.fromV2(saved);
     }
 
-    public void removeReservation(final long id) {
-        int deleteCounts = reservationRepository.deleteById(id);
-        if (deleteCounts == DELETE_FAILED_COUNT) {
-            throw new NoSuchElementException(String.format("[ERROR] 예약번호 %d번은 존재하지 않습니다.", id));
-        }
+    public ReservationResponse addReservationForAdmin(final AdminReservationRequest request) {
+        final long timeId = request.timeId();
+        final long themeId = request.themeId();
+        final Member member = memberRepository.findById(request.memberId())
+                .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 사용자 입니다."));
+        final ReservationTime time = reservationTimeRepository.findById(timeId)
+                .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 예약 시간 입니다."));
+        final ReservationTheme theme = reservationThemeRepository.findById(themeId)
+                .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 테마 입니다."));
+        final LocalDate date = request.date();
+        validateDuplicateReservation(date, timeId, themeId);
+        final Reservation reservation = Reservation.builder()
+                .member(member)
+                .date(date)
+                .time(time)
+                .theme(theme)
+                .build();
+        Reservation saved = reservationRepository.saveWithMember(reservation);
+        return ReservationResponse.fromV2(saved);
     }
 
-    private void validateUniqueReservation(final Reservation reservation) {
-        if (existsSameReservation(reservation)) {
-            throw new IllegalArgumentException("[ERROR] 이미 존재하는 예약시간입니다.");
-        }
+    public List<ReservationResponse> getAllReservations() {
+        return reservationRepository.findAllReservationsV2().stream()
+                .map(ReservationResponse::fromV2)
+                .toList();
     }
 
-    private boolean existsSameReservation(final Reservation reservation) {
-        List<Reservation> reservations = reservationRepository.findByDate(reservation.getDate());
+    public List<ReservationResponse> getFilteredReservations(final Long memberId, final Long themeId,
+                                                             final LocalDate dateFrom, final LocalDate dateTo) {
+        final List<Reservation> reservations = reservationRepository.findByMemberIdAndThemeIdAndDateFromAndDateTo(
+                memberId, themeId, dateFrom, dateTo);
         return reservations.stream()
-                .anyMatch(candidate -> candidate.isDuplicateReservation(reservation));
+                .map(ReservationResponse::fromV2)
+                .toList();
+    }
+
+    private void validateDuplicateReservation(final LocalDate localDate, final long timeId, final long themeId) {
+        if (reservationRepository.existByDateAndTimeIdAndThemeId(localDate, timeId, themeId)) {
+            throw new IllegalArgumentException("[ERROR] 이미 존재하는 예약 입니다.");
+        }
+    }
+
+    public List<MyPageReservationResponse> getReservationsByMemberId(Long memberId) {
+        final Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 사용자 입니다."));
+        List<Reservation> myReservations = reservationRepository.findByMemberId(member.getId());
+        return myReservations.stream()
+                .map(MyPageReservationResponse::from)
+                .toList();
+    }
+
+    public void removeReservation(Long reservationId) {
+        final Reservation reservation = reservationRepository.findById(reservationId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 예약입니다."));
+        reservationRepository.deleteById(reservation.getId());
     }
 }
