+++ src/test/java/roomescape/MissionStepTest.java
@@ -20,8 +20,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import roomescape.controller.ReservationController;
-import roomescape.dto.LoginRequest;
-import roomescape.dto.ReservationResponse;
+import roomescape.dto.request.LoginRequest;
+import roomescape.dto.response.ReservationResponse;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD)
