+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -1,72 +1,177 @@
 package roomescape.service;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import java.util.List;
 import java.util.NoSuchElementException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
+import org.springframework.test.annotation.DirtiesContext.ClassMode;
+import org.springframework.test.context.TestPropertySource;
+import roomescape.dto.request.MemberRegisterRequest;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.request.ReservationThemeRequest;
+import roomescape.dto.request.ReservationTimeRequest;
+import roomescape.dto.response.ReservationResponse;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+@SpringBootTest
+@DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
+@TestPropertySource(properties = {
+        "spring.sql.init.mode=never",
+        "spring.jpa.hibernate.ddl-auto=create-drop"
+})
 class ReservationServiceTest {
 
     @Autowired
-    ReservationService service;
+    private ReservationService reservationService;
 
     @Autowired
-    ReservationTimeService timeService;
+    private MemberService memberService;
 
-    @DisplayName("같은 날짜 및 시간 예약이 존재하면 예외를 던진다")
+    @Autowired
+    private ReservationTimeService reservationTimeService;
+
+    @Autowired
+    private ReservationThemeService reservationThemeService;
+
     @Test
-    void addReservationWithDuplicatedReservation() {
-        //given
-        LocalDate date = LocalDate.now().plusDays(1);
-        
-        ReservationTimeResponse response = timeService.addReservationTime(
-                new ReservationTimeRequest(LocalTime.parse("10:00")));
-
-        service.addReservation(new ReservationRequest("test", date, 1L, response.id()));
-
-        //when & then
-        ReservationRequest duplicated = new ReservationRequest("test2", date, 1L, response.id());
-        assertThatThrownBy(() -> service.addReservation(duplicated))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 이미 존재하는 예약시간입니다.");
-
-    }
-
-    @DisplayName("현재 시점 이전의 예약을 생성할 시 예외를 던진다")
-    @Test
-    void addReservationBeforeCurrentDateTime() {
+    @DisplayName("사용자의 id를 이용해 예약을 생성한다")
+    void createReservationTest() {
         // given
-        timeService.addReservationTime(new ReservationTimeRequest(LocalTime.parse("10:10")));
-        LocalDate date = LocalDate.now().minusDays(1);
-        ReservationRequest request = new ReservationRequest("호떡", date, 1L, 1L);
+        Long memberId = memberService.addMember(new MemberRegisterRequest("", "", "")).id();
+        Long timeId = reservationTimeService.addReservationTime(new ReservationTimeRequest(LocalTime.now())).id();
+        Long themeId = reservationThemeService.addReservationTheme(new ReservationThemeRequest("", "", "")).id();
 
-        // then & when
-        assertThatThrownBy(() -> service.addReservation(request))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 이전 시각으로 예약할 수 없습니다.");
+        ReservationRequest reservationRequest = new ReservationRequest(
+                LocalDate.now().plusDays(1),
+                themeId,
+                timeId
+        );
+
+        // when
+        ReservationResponse reservationResponse = reservationService.addReservationWithMemberId(reservationRequest,
+                memberId);
+
+        // then
+        assertAll(
+                () -> assertThat(reservationResponse.id()),
+                () -> assertThat(reservationResponse.date()),
+                () -> assertThat(reservationResponse.name())
+        );
     }
 
-    @DisplayName("존재하지 않는 예약을 삭제하려는 경우 예외를 던진다")
     @Test
-    void removeReservation() {
-        //given
-        long notExistId = 999;
+    @DisplayName("사용자가 없는 theme id를 이용해 예약을 생성한다")
+    void createReservationTest2() {
+        // given
+        Long memberId = memberService.addMember(new MemberRegisterRequest("", "", "")).id();
+        Long timeId = reservationTimeService.addReservationTime(new ReservationTimeRequest(LocalTime.now())).id();
+        Long themeId = reservationThemeService.addReservationTheme(new ReservationThemeRequest("", "", "")).id();
+        Long nonExistTimeId = 2L;
+        Long nonExistThemeId = 2L;
 
-        //when & then
-        assertThatThrownBy(() -> service.removeReservation(notExistId))
-                .isInstanceOf(NoSuchElementException.class)
-                .hasMessage("[ERROR] 예약번호 999번은 존재하지 않습니다.");
+        ReservationRequest reservationRequest1 = new ReservationRequest(
+                LocalDate.now().plusDays(1),
+                nonExistThemeId,
+                timeId
+        );
+
+        ReservationRequest reservationRequest2 = new ReservationRequest(
+                LocalDate.now().plusDays(1),
+                themeId,
+                nonExistTimeId
+        );
+
+        // when, then
+        assertAll(
+                () -> assertThatThrownBy(
+                        () -> reservationService.addReservationWithMemberId(reservationRequest1, memberId)
+                ).isInstanceOf(NoSuchElementException.class),
+                () -> assertThatThrownBy(
+                        () -> reservationService.addReservationWithMemberId(reservationRequest2, memberId)
+                ).isInstanceOf(NoSuchElementException.class)
+        );
     }
 
+    @Test
+    @DisplayName("미래가 아닌 날짜로 예약 시도 시 예외 발생")
+    void createReservationTest3() {
+        // given
+        Long memberId = memberService.addMember(new MemberRegisterRequest("", "", "")).id();
+        Long timeId = reservationTimeService.addReservationTime(new ReservationTimeRequest(LocalTime.now())).id();
+        Long themeId = reservationThemeService.addReservationTheme(new ReservationThemeRequest("", "", "")).id();
+
+        ReservationRequest reservationRequest1 = new ReservationRequest(
+                LocalDate.now(),
+                themeId,
+                timeId
+        );
+
+        ReservationRequest reservationRequest2 = new ReservationRequest(
+                LocalDate.now().minusDays(1),
+                themeId,
+                timeId
+        );
+
+        // when, then
+        assertAll(
+                () -> assertThatThrownBy(
+                        () -> reservationService.addReservationWithMemberId(reservationRequest1, memberId)
+                ).isInstanceOf(IllegalArgumentException.class),
+                () -> assertThatThrownBy(
+                        () -> reservationService.addReservationWithMemberId(reservationRequest2, memberId)
+                ).isInstanceOf(IllegalArgumentException.class)
+        );
+    }
+
+
+    @DisplayName("예약이 중복되어 예외가 발생 한다.")
+    @Test
+    void duplicateTest() {
+        // given
+        Long memberId = memberService.addMember(new MemberRegisterRequest("", "", "")).id();
+        Long timeId = reservationTimeService.addReservationTime(new ReservationTimeRequest(LocalTime.now())).id();
+        Long themeId = reservationThemeService.addReservationTheme(new ReservationThemeRequest("", "", "")).id();
+
+        ReservationRequest reservationRequest = new ReservationRequest(
+                LocalDate.now().plusDays(1),
+                themeId,
+                timeId
+        );
+        reservationService.addReservationWithMemberId(reservationRequest,
+                memberId);
+
+        // when & then
+        assertThatThrownBy(() -> reservationService.addReservationWithMemberId(reservationRequest, memberId))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    @DisplayName("모든 예약 정보를 가져온다.")
+    void getAllReservationsTest() {
+        // given
+        Long memberId = memberService.addMember(new MemberRegisterRequest("", "", "")).id();
+        Long timeId = reservationTimeService.addReservationTime(new ReservationTimeRequest(LocalTime.now())).id();
+        Long themeId = reservationThemeService.addReservationTheme(new ReservationThemeRequest("", "", "")).id();
+
+        ReservationRequest reservationRequest = new ReservationRequest(
+                LocalDate.now().plusDays(1),
+                themeId,
+                timeId
+        );
+        reservationService.addReservationWithMemberId(reservationRequest,
+                memberId);
+        //when
+        final List<ReservationResponse> expected = reservationService.getAllReservations();
+
+        //then
+        assertThat(expected).hasSize(1);
+    }
 }
